+++ Repository/AakSamlTeamMetaRepository.php
@@ -1,5 +1,13 @@
 <?php
 
+/*
+ * This file is part of the "AakSamlBundle" for Kimai.
+ * All rights reserved by ITK Development (https://github.com/itk-kimai).
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace KimaiPlugin\AakSamlBundle\Repository;
 
 use Doctrine\ORM\EntityRepository;
@@ -12,8 +20,6 @@ use KimaiPlugin\AakSamlBundle\Entity\AakSamlTeamMeta;
 class AakSamlTeamMetaRepository extends EntityRepository
 {
     /**
-     * @param AakSamlTeamMeta $meta
-     *
      * @throws ORMException
      * @throws \Doctrine\ORM\OptimisticLockException
      */